FIX: Further fix to check resetting
The Check objects need to be recreated as they also have internal state.

As an aside, putting caching / persistent data through all the layers
of an app makes these kind of bugs more necessary. It can be helpful,
instead, to keep most internal layers of a system quite “functional”
and not persist state, and then deal with whatever caching/persistence
is needed to get a performant result only on the outermost layers.